Address review feedback for quote-and-bind
- premium modelled as decimal string end-to-end (BigDecimal/Java, string/TS)
- POST /api/policies is idempotent: returns 200 on replay, 201 on first bind;
  Policy.quoteId is unique-indexed
- QuoteRequest carries bean-validation constraints (pattern/size); future DOB
  rejected in EligibilityService; MethodArgumentNotValidException -> standard
  ErrorResponse
- Clock bean injected into Eligibility/Rating/PolicyService for deterministic
  time-coupled logic
- App.tsx: separate bindError state shown inside quote panel; quote panel
  hides on successful bind; bind retry remains available on failure

From: https://github.com/nrussell94/ai-delivery-demo/pull/2#pullrequestreview-4305831857

+++ backend/src/main/java/com/example/demo/service/QuoteService.java
@@ -40,7 +40,7 @@ public class QuoteService {
 
         QuoteResponse response = new QuoteResponse();
         response.setQuoteId(saved.getId());
-        response.setPremium(saved.getPremium().doubleValue());
+        response.setPremium(saved.getPremium());
         return response;
     }
 }
